Refactored so API is now its own module
The XACML editor is just a small module that uses the XACML API. There
are use cases where people will want to use the XACML API but not enable
the islandora xacml editor. I moved all the API files to a directory
called API and created a new module that the editor module depends
on called islandora_xacml_api.