Islandora XACML editor failing on root collections
Colorado Alliance was having a problem were the XACML editor was
failing very badly on top level collections. It turns out that the
case where a PID was not being explicitly passed in the URL wasn't
being handled correctly. This is now fixed.

This update also makes sure if you run a batch on a collection with
no children that a divide by zero is not done, which ends up in a
infinite loop until memory ran out. This should never happen under
normal circumstances, but its good to have the check there anyway.

It also added some more testing to make sure that pids that are not updated
are shown to the user.