Added optional argurment to constructFromPid
Xacml::constructFromPid used to only take one arguement, the PID to work
with. I added an optional arguement so that a DSID can also be passed. My
use case is specifically to parse the CHILD_SECURITY stream, but I think it
is a useful general function to have. Its an optional arguemnt so when its
not specified 'POLICY' will still be used.