Modified XACML so administrator role isn't special
The administrator and fedoraAdmin roles used to be treated as special
they would be hidden from view when the getUser and getRoles functions
were called, because they were assumed to be added by the library (for
safety they are always included in the policy).

This causes problems when the administrator role is specifically added.
Users wonder why it is missing. So we now make sure that every policy
includes fedoraAdmin and administrator, but we don't hide that.