Updated how Object is stored in XACML forms
Before the object was being stored in the form state. When the object
was unserialized in the submit function it was a different object
then the one loaded by islandora_load_object. This lead to inconsistancy
when the RELS-INT was updated on one object but not the other, since
they had seperate caches.

This was fixed in this commit, by passing the pid to be serialized
and then reloading the object using islandora_object_load when it
is needed, so we have a consistant copy in all places.

This may lead to a discussion about how we implement caching, but
this will fix the problem while we think about a larger solution.